Move git.Diff to gitserver.Client
